# docs/openapi.yaml
openapi: 3.0.3
info:
  title: DataFlow API
  description: Secure, self-hosted family data vault API
  version: 0.2.0
  license:
    name: MIT
servers:
  - url: http://localhost:3460
    description: Development server
paths:
...
  /api/auth/register:
    post:
      summary: Register a new user
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, master_password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                master_password:
                  type: string
                display_name:
                  type: string
      responses:
        '201':
          description: User registered
  /api/auth/login:
    post:
      summary: Log in
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, master_password]
              properties:
                email:
                  type: string
                password:
                  type: string
                master_password:
                  type: string
      responses:
        '200':
          description: Login successful
...
  /api/auth/sessions/{sid}:
    delete:
      summary: Revoke a session
      tags: [Auth]
      parameters:
        - name: sid
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Session revoked
...
  /api/data/import:
    post:
      summary: Import data
      tags: [Data]
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                format:
                  type: string
                  enum: [bitwarden, chrome, lastpass, onepassword, keepass]
      responses:
        '200':
          description: Import result
  /api/data/export:
    get:
      summary: Export vault data
      tags: [Data]
      parameters:
        - name: format
          in: query
          schema:
            type: string
            enum: [json, csv, pdf]
      responses:
        '200':
          description: Exported data
...